Accept cost amounts in trips controller
Since it doesn't make sense for a trip to have no requested costs and
vice versa the trip controller should create these records in a
transaction. There was an error running a transaction within the testing
environment while accessing the db from process.db, so this global was
removed entirely.

The stated error was that this.client was undefined inside the Query
postgres dialect instance. So the connection was being lost somehow.
Removing the process.db.sequelize.close() call in globalTeardown didn't
fix this issue.