Prevent email body styles from affecting panels
Opening old email logs caused the text of the expansion panel title to
blow up to 1.5em. This was because the older email logs use body styles
that ended up applying to the whole page. The new emails do not have
this problem. We can apply a small workaround by fixing the height of
the expansion panel text.