Set trip fiscal year to largest overlap
There are many trips with a departure date of June 30th, 2018. Phyllis
asked for these to be FY2019 trips rather than FY2018. We determined
that a ruling where travel requests are categorized by the maxiumum
number of days they fall into would work better than going off of just
the departure date.

Ex: 6/30/18-7/5/18 would be FY19 since there are 5 days in FY19 compared
to the 1 day in FY18.