Incorporate balance in getSeniorFundsLeft
It's possible for the result of funds allocated for seniors subtracted
by the funds used by seniors to be greater than the balance of a budget
allocation. A simple example would be when only juniors have used those
funds. In this case the expression described above would just be the
original amount reserved for seniors.

This doesn't make a lot of sense in the context of "getSeniorFundsLeft".
What we really want is the budget balance to be returned if it's less
than the funds seniors can still use.

This allows us to simplify the frontend a bit which was doing this MIN
operation anyway.