Prevent users from submitting non-open requests
This is only a frontend block at the moment. A user could theoretically
submit directly to the POST endpoint and create a new travel request
outside of these ranges.

I'm not going to do a backend block. The funding controller can deny the
request manually if one slips through through this mechanism.