Fix error in Trip.fullExport
I received a bug report from a user that the export was not working.
I assume the error was related to the recent updates which changed
parts of the data model. After looking over those recent changes,
I came to understand these changes as being the appropriate fixes.

Since I have not worked on this project before, I did not know how
to verify myself if the exported data was correct after these fixes.
I sent an export file to the user who reported the issue to check
if the report looked correct to them. They said it looked correct,
so we will be rolling out this fix.

In the future, I'd like to find a way to write an automated test to
verify the correctness of the export output so changes to this
feature can be made with higher confidence.