Fix error in calculating "Funds Left for Seniors".
It was reported that the calculation was returning the same value
for "Funds Left for Seniors" as "Budget Balances" which raised suspicion
that the calculation for "Funds Left for Seniors" was incorrect since
it should be lower than the total budget.

After some investigation, I believe I found the root cause in
BudgetAllocation.prototype.getSeniorFundsLeft - when calculating
the amount of $ granted to "senior" faculty (getSeniorGrantsTotal)
the query was only counting Grants from the current BudgetAllocation.
However, the desired logic is to count the Grants from all BudgetAllocations
within the parent Budget.

In other words, when we are calculating the "senior" funds left for
Spring FY 20, we want to consider all of the senior Grants awarded
for both Spring FY 20 and Fall FY 20. The previous version of this
query would only include the Grants from Spring FY 20. This change
makes it include both, which should fix the calculation.

It is difficult for me to confirm this new value is exactly correct,
so I can't completely rule out other problems at hand in this calculation.
Based on my discussions with the admins and some tests, I belive this
is at least a step in the right direction so I plan to release it for
them to evaluate.