Fix bug causing "Standing" column to always have "junior" in the export report.
Since the trip.isForSenior function relies on startDate, this check was
working incorrectly when startDate was renamed via "as" in SQL.

This fix feels a bit hacky, but due to the importance of fixing the
functionality for the admins users I am going to use it for now. Ideally,
I would like to refactor this function to separate the logic of the
export (e.g., the SQL query) and the presentation of the repot data
(e.g., the name, order, etc. of columns in the download file).